test: move all KV related RESTful APIs into a separate test file
